✅ Chore: Service Taget,  API 하나의 파일로 합침 (#96)
